+++ src/App.js
@@ -1,29 +1,13 @@
 import React from 'react';
-
-let todoList = [
-  {
-    id: 1,
-    title: 'Complete assignment'
-  },
-  {
-    id: 2,
-    title: 'Read the book'
-  },
-  {
-    id: 3,
-    title: 'Ask questions'
-  }
-]
+import AddTodoForm from './AddTodoForm';
+import TodoList from './TodoList';
 
 function App() {
   return (
     <div>
       <h1>Todo List</h1>
-      <ul>
-        {todoList.map(function(item) {
-          return <li key={item.id}>{item.title}</li>
-        })}
-      </ul>
+        <AddTodoForm />
+        <TodoList />
     </div>
   );
 }
